Move opencl.hpp to hpxcl/opencl.hpp and change all includes
