+++ examples/opencl/list_devices.cpp
@@ -7,7 +7,7 @@
 #include <hpx/include/iostreams.hpp>
 #include <hpx/lcos/future.hpp>
 
-#include "../../opencl.hpp"
+#include <hpxcl/opencl.hpp>
 #include <sstream>
 
 using namespace hpx::opencl;
